+++ RELEASE_NOTES.md
@@ -28,8 +28,10 @@
 * [IOTDB-1466] Support device template
 * [IOTDB-1491] UDTF query supported in cluster
 * [IOTDB-1536] Support fuzzy query
+* [IOTDB-1496] Timed flush memtable
 * [IOTDB-1561] Support fill by specific value
 * [IOTDB-1565] Add sql: set system to readonly/writable
+* [IOTDB-1569] Timed close TsFileProcessor
 * TTL can be set to the prefix path of storage group
 * add JMX monitor to all ThreadPools in the server module